move types in crate::types to crate::server
